+++ pkg/runnerfiles/common.go
@@ -30,7 +30,7 @@ func appendString(path string, value string) error {
 		return err
 	}
 
-	file, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o644)
+	file, err := os.OpenFile(path, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o600)
 	if err != nil {
 		return errors.Wrapf(err, "open runner file %s", path)
 	}